yaffs-direct: Add reldir and reldev interfaces
These interfaces allow far more flexibility in the way Yaffs Direct
is hooked up to an RTOS VFS.

The "normal" interface requires a fully, qualified path, which is
a pain to achieve in some RTOSs.

The new interfaces allow calls to be done with paths relative to a
directory or to a device. This tends to make some RTOS integration
cleaner.

Example
 yaffs_open("/mountpoint-name/dir/file",...);
now has two alternatives:
 yaffs_open_reldir(root_dir,"dir/file",...);
or
 yaffs_open_reldev(dev, "dir/file",...);

Signed-off-by: Charles Manning <cdhmanning@gmail.com>