Fix /proc/yaffs state dumping when YAFFS_NEW_PROCFS is defined
Some of the state of Yaffs mounts can be determined by reading the /proc/yaffs entry:
  - https://yaffs.net/documents/yaffs-tuning

That was not working when YAFFS_NEW_PROCFS is defined because in order to reuse
the same routine for older proc APIs and newer ones, the function was only being called
with a fixed step parameter instead of iterating into it while it still has data. Current
version reuses the original function but iterates on it.